Cleanup src/vs/platform/environment/node/argv.ts. Fixes #76239
