+++ src/vs/code/test/node/argv.test.ts
@@ -3,7 +3,8 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 import * as assert from 'assert';
-import { formatOptions, Option, addArg } from 'vs/platform/environment/node/argv';
+import { formatOptions, Option } from 'vs/platform/environment/node/argv';
+import { addArg } from 'vs/platform/environment/node/argvHelper';
 
 suite('formatOptions', () => {
